Move logic into a command-helper
