+++ src/commands/build.js
@@ -1,8 +1,8 @@
 const chalk = require('chalk')
-const yaml = require('yaml');
 
 const { createCompiler } = require('../command-helpers/compiler')
 const { fixParameters } = require('../command-helpers/gluegun')
+const { updateSubgraphNetwork } = require('../command-helpers/networks')
 const DataSourcesExtractor = require('../command-helpers/data-sources')
 const Protocol = require('../protocols')
 
@@ -81,49 +81,18 @@ module.exports = {
     }
 
 
+/// NETWORK FILE
     networksFile = networksFile || "./networks.json"
 
     if (network && !filesystem.exists(networksFile)) {
-      print.error(`Network file '${networksFile}' does not exists!`)
-      process.exitCode = 1
-      return
-    } else if (network) {
-      let networksObj = filesystem.read(networksFile, "json")
-      let networkObj = networksObj[network]
-
-      if(!networkObj) {
-        print.error(`Could not find network with name '${network}' in '${networksFile}'`)
-        process.exitCode = 1
-        return
-      }
-
-      await patching.update(manifest, subgraph => {
-        let doc = yaml.parse(subgraph)
-        let networkSources = Object.keys(networkObj)
-
-        doc["dataSources"].forEach( ds => {
-            if (!networkSources.includes(ds.name)) {
-              print.info(`Skipping ${ds.name} - not in networks config`)
-              return
-            }
-
-            let dsNetwork = networkObj[ds.name]
-            if (hasChanges(network, dsNetwork, ds)) {
-              ds.network = network
-              ds.source = { abi: ds.source.abi }
-              Object.assign(ds.source, dsNetwork)
-            } else {
-              print.info(`Skipping ${ds.name} - no changes`)
-            }
-        })
-
-        let yaml_doc = new yaml.Document();
-        yaml_doc.contents = doc
-        return yaml_doc.toString()
-      })
+    print.error(`Network file '${networksFile}' does not exists!`)
+    process.exitCode = 1
+    return
     }
 
-    return
+    if (network) {
+      await updateSubgraphNetwork(manifest, network, networksFile)
+    }
 
     let protocol
     try {
@@ -161,20 +130,4 @@ module.exports = {
       }
     }
   },
-}
-
-function hasChanges(network, networkObj, dataSource) {
-  let networkChanged = dataSource.network !== network
-
-  let addressChanged
-
-  if (network === "near") {
-    addressChanged = networkObj.account !== dataSource.source.account
-  } else {
-    addressChanged = networkObj.address !== dataSource.source.address
-  }
-
-  let startBlockChanged = networkObj.startBlock !== dataSource.source.startBlock
-
-  return networkChanged || addressChanged || startBlockChanged
 }
